+++ src/wl_measures/wl_measures_bayes_factor.py
@@ -30,24 +30,24 @@ def to_freqs_sections_tokens(main, tokens, tokens_x1, tokens_x2, measure_bayes_f
     freqs_sections_tokens = {}
 
     if measure_bayes_factor == _tr('wl_measures_bayes_factor', "Student's t-test (2-sample)"):
-        num_sections = main.settings_custom['measures']['bayes_factor']['students_t_test_2_sample']['num_sections']
+        num_sub_sections = main.settings_custom['measures']['bayes_factor']['students_t_test_2_sample']['num_sub_sections']
         use_data = main.settings_custom['measures']['bayes_factor']['students_t_test_2_sample']['use_data']
 
-    sections_x1 = wl_nlp_utils.to_sections(tokens_x1, num_sections)
-    sections_x2 = wl_nlp_utils.to_sections(tokens_x2, num_sections)
+    sections_x1 = wl_nlp_utils.to_sections(tokens_x1, num_sub_sections)
+    sections_x2 = wl_nlp_utils.to_sections(tokens_x2, num_sub_sections)
 
-    sections_freqs_x1 = [collections.Counter(section) for section in sections_x1]
-    sections_freqs_x2 = [collections.Counter(section) for section in sections_x2]
+    freqs_sections_x1 = [collections.Counter(section) for section in sections_x1]
+    freqs_sections_x2 = [collections.Counter(section) for section in sections_x2]
 
     if use_data == _tr('wl_measures_bayes_factor', 'Absolute Frequency'):
         for token in tokens:
             freqs_x1 = [
-                section_freqs.get(token, 0)
-                for section_freqs in sections_freqs_x1
+                freqs_section.get(token, 0)
+                for freqs_section in freqs_sections_x1
             ]
             freqs_x2 = [
-                section_freqs.get(token, 0)
-                for section_freqs in sections_freqs_x2
+                freqs_section.get(token, 0)
+                for freqs_section in freqs_sections_x2
             ]
 
             freqs_sections_tokens[token] = (freqs_x1, freqs_x2)
@@ -57,12 +57,12 @@ def to_freqs_sections_tokens(main, tokens, tokens_x1, tokens_x2, measure_bayes_f
 
         for token in tokens:
             freqs_x1 = [
-                section_freqs.get(token, 0) / len_section
-                for section_freqs, len_section in zip(sections_freqs_x1, len_sections_x1)
+                freqs_section.get(token, 0) / len_section
+                for freqs_section, len_section in zip(freqs_sections_x1, len_sections_x1)
             ]
             freqs_x2 = [
-                section_freqs.get(token, 0) / len_section
-                for section_freqs, len_section in zip(sections_freqs_x2, len_sections_x2)
+                freqs_section.get(token, 0) / len_section
+                for freqs_section, len_section in zip(freqs_sections_x2, len_sections_x2)
             ]
 
             freqs_sections_tokens[token] = (freqs_x1, freqs_x2)
@@ -72,15 +72,20 @@ def to_freqs_sections_tokens(main, tokens, tokens_x1, tokens_x2, measure_bayes_f
 # Log-likelihood Ratio
 # Reference: Wilson, A. (2013). Embracing Bayes Factors for key item analysis in corpus linguistics. In M. Bieswanger, & A. Koll-Stobbe (Eds.), New Approaches to the Study of Linguistic Variability (pp. 3–11). Peter Lang.
 def bayes_factor_log_likelihood_ratio_test(main, c11, c12, c21, c22):
+    cxx = c11 + c12 + c21 + c22
+
     log_likelihood_ratio = wl_measures_statistical_significance.log_likelihood_ratio_test(main, c11, c12, c21, c22)[0]
-    bic = log_likelihood_ratio - numpy.log(c11 + c12 + c21 + c22)
+    bic = log_likelihood_ratio - numpy.log(cxx) if cxx else 0
 
     return bic
 
 # Student's t-test (2-sample)
 # Reference: Wilson, A. (2013). Embracing Bayes Factors for key item analysis in corpus linguistics. In M. Bieswanger, & A. Koll-Stobbe (Eds.), New Approaches to the Study of Linguistic Variability (pp. 3–11). Peter Lang.
-def bayes_factor_students_t_test_2_sample(main, counts_x1, counts_x2):
-    t_stat = wl_measures_statistical_significance.students_t_test_2_sample(main, counts_x1, counts_x2)[0]
-    bic = t_stat ** 2 - numpy.log(2 * main.settings_custom['measures']['bayes_factor']['students_t_test_2_sample']['num_sections'])
+def bayes_factor_students_t_test_2_sample(main, freqs_x1, freqs_x2):
+    if any(freqs_x1) or any(freqs_x2):
+        t_stat = wl_measures_statistical_significance.students_t_test_2_sample(main, freqs_x1, freqs_x2)[0]
+        bic = t_stat ** 2 - numpy.log(2 * main.settings_custom['measures']['bayes_factor']['students_t_test_2_sample']['num_sub_sections'])
+    else:
+        bic = 0
 
     return bic