+++ src/wl_tests/wl_tests_work_area/test_ngram_generator.py
@@ -47,23 +47,23 @@ def test_ngram_generator():
         random_i = random.randrange(0, 10)
 
         # Single file with search terms
-        if random_i in [0, 3, 6, 9]:
+        if random_i in [0, 2, 4, 6]:
             random.choice(files)['selected'] = True
 
             main.settings_custom['ngram_generator']['search_settings']['search_settings'] = True
         # Single file without search terms
-        elif random_i in [1, 4, 7]:
+        elif random_i == 8:
             random.choice(files)['selected'] = True
 
             main.settings_custom['ngram_generator']['search_settings']['search_settings'] = False
         # Multiple files with search terms
-        elif random_i in [2, 5]:
+        elif random_i in [1, 3, 5, 7]:
             for file in random.sample(files, 2):
                 file['selected'] = True
 
             main.settings_custom['ngram_generator']['search_settings']['search_settings'] = True
         # Multiple files without search terms
-        elif random_i == 8:
+        elif random_i == 9:
             for file in random.sample(files, 2):
                 file['selected'] = True
 
@@ -94,6 +94,7 @@ def test_ngram_generator():
     main.app.quit()
 
 def update_gui(err_msg, ngrams_freq_files, ngrams_stats_files):
+    print(err_msg)
     assert not err_msg
 
     len_files_selected = len(list(main.wl_file_area.get_selected_files()))